Правка названий классов
